+++ NEWS.md
@@ -42,6 +42,7 @@ None
 35. Fix #82 - Add function `util_logistic_param_estimate()`
 36. Fix #83 - Add function `util_negative_binomial_param_estimate()`
 37. Fix #84 - Add function `util_normal_param_estimate()`
+38. Fix #85 - Add function `util_pareto_param_estimate()`
 
 ## Fixes and Minor Improvements
 1. Fix #30 - Move `crayon`, `rstudioapi`, and `cli` from Suggests to Imports due to `pillar`